support asm feature for invert
