+++ primitive/jubjub/src/arithmetic/limbs.rs
@@ -10,4 +10,4 @@ mod normal;
 pub(crate) use assembly::{add, double, mul, neg, square, sub};
 
 #[cfg(any(not(feature = "asm"), not(target_arch = "x86_64")))]
-pub(crate) use normal::{add, double, invert, mul, neg, square, sub};
+pub(crate) use normal::{add, double, mul, neg, square, sub};
