+++ primitive/jubjub/src/domain/field.rs
@@ -135,7 +135,7 @@ macro_rules! field_operation {
             }
 
             pub fn invert(self) -> Option<$field> {
-                invert(&self).map(|x| $field(x))
+                Self::_invert(&self).map(|x| $field(x))
             }
 
             pub fn random(mut rand: impl RngCore) -> $field {
@@ -149,6 +149,113 @@ macro_rules! field_operation {
                     rand.next_u64(),
                     rand.next_u64(),
                 ])
+            }
+
+            #[inline]
+            fn _invert(x: &$field) -> Option<[u64; 4]> {
+                #[inline(always)]
+                fn square_assign_multi(n: &mut Fr, num_times: usize) {
+                    for _ in 0..num_times {
+                        *n = n.square();
+                    }
+                }
+                let mut t1 = x.square();
+                let mut t0 = t1.square();
+                let mut t3 = t0 * t1;
+                let t6 = t3 * *x;
+                let t7 = t6 * t1;
+                let t12 = t7 * t3;
+                let t13 = t12 * t0;
+                let t16 = t12 * t3;
+                let t2 = t13 * t3;
+                let t15 = t16 * t3;
+                let t19 = t2 * t0;
+                let t9 = t15 * t3;
+                let t18 = t9 * t3;
+                let t14 = t18 * t1;
+                let t4 = t18 * t0;
+                let t8 = t18 * t3;
+                let t17 = t14 * t3;
+                let t11 = t8 * t3;
+                t1 = t17 * t3;
+                let t5 = t11 * t3;
+                t3 = t5 * t0;
+                t0 = t5.square();
+                square_assign_multi(&mut t0, 5);
+                t0 *= t3;
+                square_assign_multi(&mut t0, 6);
+                t0 *= t8;
+                square_assign_multi(&mut t0, 7);
+                t0 *= t19;
+                square_assign_multi(&mut t0, 6);
+                t0 *= t13;
+                square_assign_multi(&mut t0, 8);
+                t0 *= t14;
+                square_assign_multi(&mut t0, 6);
+                t0 *= t18;
+                square_assign_multi(&mut t0, 7);
+                t0 *= t17;
+                square_assign_multi(&mut t0, 5);
+                t0 *= t16;
+                square_assign_multi(&mut t0, 3);
+                t0 *= *x;
+                square_assign_multi(&mut t0, 11);
+                t0 *= t11;
+                square_assign_multi(&mut t0, 8);
+                t0 *= t5;
+                square_assign_multi(&mut t0, 5);
+                t0 *= t15;
+                square_assign_multi(&mut t0, 8);
+                t0 *= *x;
+                square_assign_multi(&mut t0, 12);
+                t0 *= t13;
+                square_assign_multi(&mut t0, 7);
+                t0 *= t9;
+                square_assign_multi(&mut t0, 5);
+                t0 *= t15;
+                square_assign_multi(&mut t0, 14);
+                t0 *= t14;
+                square_assign_multi(&mut t0, 5);
+                t0 *= t13;
+                square_assign_multi(&mut t0, 2);
+                t0 *= *x;
+                square_assign_multi(&mut t0, 6);
+                t0 *= *x;
+                square_assign_multi(&mut t0, 9);
+                t0 *= t7;
+                square_assign_multi(&mut t0, 6);
+                t0 *= t12;
+                square_assign_multi(&mut t0, 8);
+                t0 *= t11;
+                square_assign_multi(&mut t0, 3);
+                t0 *= *x;
+                square_assign_multi(&mut t0, 12);
+                t0 *= t9;
+                square_assign_multi(&mut t0, 11);
+                t0 *= t8;
+                square_assign_multi(&mut t0, 8);
+                t0 *= t7;
+                square_assign_multi(&mut t0, 4);
+                t0 *= t6;
+                square_assign_multi(&mut t0, 10);
+                t0 *= t5;
+                square_assign_multi(&mut t0, 7);
+                t0 *= t3;
+                square_assign_multi(&mut t0, 6);
+                t0 *= t4;
+                square_assign_multi(&mut t0, 7);
+                t0 *= t3;
+                square_assign_multi(&mut t0, 5);
+                t0 *= t2;
+                square_assign_multi(&mut t0, 6);
+                t0 *= t2;
+                square_assign_multi(&mut t0, 7);
+                t0 *= t1;
+
+                match x.is_zero() {
+                    true => None,
+                    false => Some(t0.0),
+                }
             }
         }
